pyproject.toml
[tool.poetry]
name = "dual-path-rnn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
librosa = "^0.7.2"
tensorflow = "^2.1.0"
attrs = "^19.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
force_single_line = true
default_section = "THIRDPARTY"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

